// frontend/src/api/model/spectralRecord.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { SpectralRecordArtifact } from './spectralRecordArtifact';
import type { File } from './file';
import type { OrganizationSummary } from './organizationSummary';
import type { UserSummary } from './userSummary';
import type { ProcessingStatusEnum } from './processingStatusEnum';

/**
 * Serializer for spectral records.
 */
export interface SpectralRecord {
  readonly id: string;
  readonly artifacts: readonly SpectralRecordArtifact[];
  readonly artifacts_count: string;
  readonly raw_file: File;
  readonly owner: OrganizationSummary;
  readonly author: UserSummary;
  is_deleted?: boolean;
  /** @nullable */
  deleted_at?: string | null;
  /** Status of async background processing

* `pending` - Pending processing
* `processing` - Processing in progress
* `completed` - Processing completed
* `failed` - Processing failed */
  processing_status?: ProcessingStatusEnum;
  /**
   * Name of this record. Short and simple description of record.
   * @maxLength 80
   * @nullable
   */
  name?: string | null;
  /** Description of the record */
  description?: string;
  /** Tick this box if the record is dependent on absolute time. When you need align record to real time. */
  time_tracked?: boolean;
  /**
   * System time of record start
   * @nullable
   */
  time_internal_start?: number | null;
  /**
   * When 'time is tracked', you can set start time of the record beginning. 
   * @nullable
   */
  time_start?: string | null;
  /** @nullable */
  time_of_interest_start?: number | null;
  /** @nullable */
  time_of_interest_end?: number | null;
  readonly created: string;
  /**
   * Duration of record
   * @nullable
   */
  record_duration?: string | null;
  /** record metadata, used for advanced data processing and maintaining */
  metadata?: unknown;
  /** @nullable */
  deleted_by?: number | null;
  /** @nullable */
  detector?: string | null;
  /** @nullable */
  calib?: string | null;
}